% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDTM.R
\name{findAssocs}
\alias{findAssocs}
\title{findAssocs}
\usage{
findAssocs(dtm, term, corlimit = 0.3)
}
\arguments{
\item{dtm}{matrix. Document Term Matrix}

\item{term}{character. keyword}

\item{corlimit}{numeric. lower correlation limit}
}
\value{
character. correlation above corlimit.
}
\description{
find association by one keyword
}
\examples{
findAssocs(dtm = dtm, term = "home", corlimit = 0.8)
}
