% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDTM.R
\name{findAssocsAll}
\alias{findAssocsAll}
\title{findAssocsAll}
\usage{
findAssocsAll(dtm, corlimit = 0.3)
}
\arguments{
\item{dtm}{matrix. Document Term Matrix}

\item{corlimit}{numeric. lower correlation limit}
}
\value{
character. correlation above corlimit.
}
\description{
find association of all terms
}
\examples{
findAssocsAll(dtm = dtm, corlimit = 0.8)
}
