% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDTM.R
\name{findAssocTwo}
\alias{findAssocTwo}
\title{findAssocTwo}
\usage{
findAssocTwo(dtm, term1, term2)
}
\arguments{
\item{dtm}{matrix. Document Term Matrix}

\item{term1}{character. keyword 1.}

\item{term2}{character. keyword 2.}
}
\value{
numeric. correlation of two terms.
}
\description{
find association of two terms
}
\examples{
findAssocTwo(dtm = dtm, term1 = "home", term2 = "family")
}
