% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDTM.R
\name{findFreqTerms}
\alias{findFreqTerms}
\title{findFreqTerms}
\usage{
findFreqTerms(dtm, lowfreq = 0, highfreq = 0)
}
\arguments{
\item{dtm}{matrix. Document Term Matrix}

\item{lowfreq}{numeric. low frequency limit}

\item{highfreq}{numeric. high frequency limit}
}
\value{
character. words which met frequency condition
}
\description{
find frequently used words(lowfreq) or below maximum frequency words(highfreq)
}
\examples{
findFreqTerms(dtm = dtm, lowfreq=3, highfreq=4)
}
