% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDTM.R
\name{makeDTM}
\alias{makeDTM}
\title{makeDTM}
\usage{
makeDTM(docs, key = "all", LABEL = FALSE, weight = "tf",
  TEXT.name = NULL, LABEL.name = NULL, RHINO = FALSE, pos = "ALL")
}
\arguments{
\item{docs}{dataframe. Every row is document.}

\item{key}{character. Vector type Keywords. if none, all words would be keywords.}

\item{LABEL}{logical. In need of LABEL column.}

\item{weight}{character. tf or tfidf}

\item{TEXT.name}{character. Column name which whould be TEXT.}

\item{LABEL.name}{character. Column name which would be LABEL.}

\item{RHINO}{logical. In need of morphological anaysis about TEXT column.}

\item{pos}{character. ALL(all Part-Of-Speech), noun(NNG, NNP, NP), verb(VV, VA, XR), NV(noun, verb), END(EC, EF), NNG, NNP, NP, NNB, VV, VA, XR, VX, EC, EF, EP}
}
\value{
matrix or dataframe(if LABEL=TRUE)
}
\description{
make DTM
}
\examples{
makeDTM(docs = docs, key = c("excel", "computer"), pos="noun")
makeDTM(docs = docs, key = c("excel", "computer"), LABEL = TRUE, weight = "tfidf", TEXT.name="body", LABEL.name="tag", pos="verb")
makeDTM(docs = docs, key = c("excel", "computer"), weight = "tfidf", TEXT.name="body", LABEL.name="tag", pos="verb")
}
